cgen: avoid forced memory leaks when not using autofree.
Without this change, the memory allocated by the hidden _vinit()
function will not be freed in _vcleanup(), which causes memory leaks
unless -autofree is active.
That means that even the hello-world example leaks memory.
That’s not critical in practice because those memory blocks are in use
during the whole program lifetime, but those leaks make others more
difficult to notice.